# user-service/openapi.json
{
  "openapi": "3.0.0",
  "info": { "title": "User Service API", "version": "1.0.0", "description": "Authentication, profile and cart APIs exposed through API Gateway." },
  "servers": [{ "url": "http://localhost:8000/api/users", "description": "API Gateway route" }],
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
  },
  "paths": {
...
    "/cart": { "get": { "summary": "Get current user's cart", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Cart data" }, "401": { "description": "Unauthorized" } } } },
    "/cart/add": { "post": { "summary": "Add item to cart", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["productId", "quantity", "color", "size"], "properties": { "productId": { "type": "string" }, "quantity": { "type": "integer", "minimum": 1 }, "color": { "type": "string" }, "size": { "type": "integer" } } } } } }, "responses": { "200": { "description": "Added" }, "400": { "description": "Invalid input" } } } },
    "/cart/update": { "put": { "summary": "Update cart item quantity", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["cartItemId", "quantity"], "properties": { "cartItemId": { "type": "integer" }, "quantity": { "type": "integer", "minimum": 1 } } } } } }, "responses": { "200": { "description": "Updated" }, "400": { "description": "Invalid input" } } } },
    "/cart/remove/{cartItemId}": { "delete": { "summary": "Remove cart item", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "cartItemId", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Removed" }, "404": { "description": "Not found" } } } },
...
  }
}